refactor: import toast directly
