+++ portal/static/components/index.js
@@ -1,1 +1,1 @@
-export { confirmationModal, xlModal } from './modal.js';export { openDrawer } from './drawer.js';export { showToast } from './toast.js';
+export { confirmationModal, xlModal } from './modal.js';export { openDrawer } from './drawer.js';
