+++ portal/static/src/components/index.js
@@ -1,3 +1,2 @@
 export { confirmationModal, xlModal } from './modal.js';
 export { openDrawer } from './drawer.js';
-export { showToast } from './toast.js';
