+++ portal/templates/document_detail.html
@@ -120,7 +120,7 @@ document.addEventListener('htmx:responseError', function (evt) {
 </script>
 {% if request.args.get('created') %}
 <script type="module">
-import { showToast } from '{{ url_for('static', filename='components/index.js') }}';
+import { showToast } from '{{ url_for('static', filename='components/toast.js') }}';
 showToast('Doküman başarıyla yüklendi ve onaya gönderildi');
 </script>
 {% endif %}
