+++ portal/templates/forms/sample.html
@@ -8,7 +8,7 @@
 </form>
 <script type="module">
   import { createInput, attachHelpText, attachValidation, attachTooltip } from '{{ url_for('static', filename='forms/index.js') }}';
-  import { showToast } from '{{ url_for('static', filename='components/index.js') }}';
+  import { showToast } from '{{ url_for('static', filename='components/toast.js') }}';
   const container = document.getElementById('form-container');
   const emailField = createInput({ id: 'email', name: 'email', label: 'Email', type: 'email', required: true, copyable: true });
   container.appendChild(emailField.wrapper);
